unify more code in eBPF
